requirements.txt
asgiref
bcrypt
certifi
charset-normalizer
Django==4.2.11
django-cors-headers
djangorestframework
djangorestframework-simplejwt
dnspython
idna
PyJWT
pymongo
python-dotenv
requests
sqlparse
typing_extensions
tzdata
urllib3

tensorflow-cpu==2.15.0
numpy<2.0.0
huggingface_hub
Pillow

gunicorn
whitenoise
